+++ src/main/java/pl/excellentapp/ekonkursy/scene/SceneConfig.java
@@ -18,7 +18,7 @@ public class SceneConfig {
     private final int height;
     private final Color backgroundColor;
     private final Color textColor;
-    private final int durationInSeconds;
+    private final double durationInSeconds;
     private final SceneMargin margin;
     private final List<SceneElement> elements;
